Crate sol_anon

Sol-Anon

Sol-Anon is a Solana program that implements an anonymous messaging system with whitelisting capabilities.

Overview

This program allows users to send messages anonymously, with special privileges for whitelisted users. It manages an inbox, message slots, and a whitelist of users.

Modules

constants: Defines constant values used throughout the program.

errors: Defines custom error types for the program.

instructions: Contains the main logic for each instruction the program can execute.

state: Defines the account structures used in the program.

utils: Contains utility functions used across the program.

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

constants

Constants

errors

Errors

instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

instructions

Instructions

program

Module representing the program.

sol_anon

The main program module containing all instruction handlers.

state

State

utils

Utils

Constants

ID_CONST

Const version of ID

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

entry

The Anchor codegen exposes a programming model where a user defines a set of methods inside of a #[program] module in a way similar to writing RPC request handlers. The macro then generates a bunch of code wrapping these user defined methods into something that can be executed on Solana.

entrypoint⚠

Safety

id

Returns the program ID

id_const

Const version of ID