Module sol_anon::instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

AddToWhitelist

Instruction.

ChangeAdmin

Instruction.

Initialize

Instruction.

ReclaimSlot

Instruction.

RemoveFromWhitelist

Instruction.

SendRegularMessage

Instruction.

SendWhitelistedMessage

Instruction.

WithdrawSurplusInboxBalance

Instruction.