Module sol_anon::instructions::inbox

Inbox Instructions

This module contains instructions related to managing the inbox.

Structs

ChangeAdmin

Accounts required for changing the admin of an inbox.

ChangeAdminBumps

InitializeInbox

Accounts required for initializing an inbox.

InitializeInboxBumps

WithdrawSurplusInboxBalance

Accounts required for withdrawing surplus balance from an inbox.

WithdrawSurplusInboxBalanceBumps

Functions

change_admin

Changes the admin of the inbox.

initialize

Initializes a new inbox.

withdraw_surplus_inbox_balance

Withdraws surplus balance from the inbox.